Fix: PreferredFoodRate Container와 Screen 구분
